Delete Auth_Controller.js
